requirements.txt
Bootstrap_Flask
Flask_CKEditor
Flask_Login
Flask-Gravatar
Flask_WTF
WTForms
Werkzeug
Flask
flask_sqlalchemy
SQLAlchemy
nltk
textblob
langdetect
flask_parameter_validation
PyJWT==2.8.0
psycopg2
flask-jwt-extended
bcrypt
python-dotenv